07.Appointment_System/requirements.txt
Django==2.2.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
Markdown==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
